Rename router service -> navigation service
