+++ web-ng/src/app/components/layer-list/layer-list.component.ts
@@ -20,7 +20,7 @@ import { Observable } from 'rxjs';
 import { map } from 'rxjs/internal/operators/map';
 import { Layer } from '../../shared/models/layer.model';
 import { List } from 'immutable';
-import { NavigationService } from '../../services/router/router.service';
+import { NavigationService } from '../../services/navigation/navigation.service';
 import { AuthService } from '../../services/auth/auth.service';
 
 @Component({
